Refactor receipt conversion route and models
This was necessary to clarify what is going on in the intermediary steps.
Before the change it was a bit difficult to find out what was happening.
The model files needed to be updated as well to reflect what was
necessary to get the models working. I would *HIGHLY* recommend manually
updating models moving forward and only using the cli for migrations.